style: mcu budget styles
