--- gabarito-3.py
@@ -1,46 +1,0 @@
-from dash import Dash, html, dcc, Input, Output
-import plotly.express as px
-import pandas as pd
-
-app = Dash(__name__)
-
-# assume you have a "long-form" data frame
-# see https://plotly.com/python/px-arguments/ for more options
-df = pd.read_excel("Vendas-2.xlsx")
-
-# criando o gráfico
-fig = px.bar(df, x="Produto", y="Quantidade", color="ID Loja", barmode="group")
-opcoes = list(df['ID Loja'].unique())
-opcoes.append("Todas as Lojas")
-
-
-app.layout = html.Div(children=[
-    html.H1(children='Faturamento Total dos Filmes por Fase'),
-    html.H2(children='No dropdown você poderá ver o custo individual dos filmes de cada fase'),
-    html.Div(children='''
-        Obs: Esse gráfico mostra o custo individual por fase.
-    '''),
-
-    dcc.Dropdown(opcoes, value='Todas as Lojas', id='lista_lojas'),
-
-    dcc.Graph(
-        id='grafico_quantidade_vendas',
-        figure=fig
-    )
-])
-
-@app.callback(
-    Output('grafico_quantidade_vendas', 'figure'),
-    Input('lista_lojas', 'value')
-)
-def update_output(value):
-    if value == "Todas as Lojas":
-        fig = px.bar(df, x="Produto", y="Quantidade", color="ID Loja", barmode="group")
-    else:
-        tabela_filtrada = df.loc[df['ID Loja']==value, :]
-        fig = px.bar(tabela_filtrada, x="Produto", y="Quantidade", color="ID Loja", barmode="group")
-    return fig
-
-
-if __name__ == '__main__':
-    app.run_server(debug=True)
